requirements/development.txt
sanic==20.6.3
alembic==1.4.2
gino==1.0.1
psycopg2-binary==2.8.5
loguru==0.5.2
click==7.1.2
schematics==2.1.0
pytest-sanic==1.6.1
sqlalchemy_utils==0.36.8
